Security

+++ src/main/java/com/sda/inTeams/controller/CommentController.java
@@ -2,19 +2,19 @@ package com.sda.inTeams.controller;
 
 import com.sda.inTeams.exception.InvalidOperation;
 import com.sda.inTeams.model.Comment.Comment;
-import com.sda.inTeams.model.Task.Task;
-import com.sda.inTeams.model.Task.TaskStatus;
 import com.sda.inTeams.model.User.User;
 import com.sda.inTeams.repository.TaskRepository;
 import com.sda.inTeams.repository.UserRepository;
+import com.sda.inTeams.service.AuthorizationService;
 import com.sda.inTeams.service.CommentService;
 import com.sda.inTeams.service.TaskService;
+import com.sda.inTeams.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import java.security.Principal;
 
 @Controller
 @RequestMapping("/comment")
@@ -22,78 +22,71 @@ import java.util.List;
 public class CommentController {
 
     private final CommentService commentService;
-    private final TaskRepository taskRepository;
-    private final UserRepository userRepository;
-
-    @GetMapping("/all")
-    public String getAllComments(Model model) {
-        model.addAttribute("commentList", commentService.getAll());
-        return "comment-list";
-    }
-
-    @GetMapping()
-    public String getTaskComments(Model model, @RequestParam(name = "taskId") long taskId) {
-        try {
-            model.addAttribute("commentList", commentService.getAllByTask(taskId));
-        } catch (InvalidOperation invalidOperation) {
-            invalidOperation.printStackTrace();
-            model.addAttribute("commentList", commentService.getAll());
-        }
-        return "comment-list";
-    }
+    private final TaskService taskService;
+    private final UserService userService;
+    private final AuthorizationService authorizationService;
 
     @GetMapping("/add")
-    public String addCommentForm(Model model, long taskId, long userId) {
+    public String addCommentForm(Model model, Principal principal, long taskId, long userId) {
         try {
-            taskRepository.findById(taskId).orElseThrow(() -> new InvalidOperation("Task not found!"));
-            userRepository.findById(userId).orElseThrow(() -> new InvalidOperation("User not found!"));
+            User user = authorizationService.getUserCredentials(principal).orElseThrow();
+            taskService.getByIdOrThrow(taskId);
+            userService.getByIdOrThrow(userId);
             model.addAttribute("newComment", new Comment());
             model.addAttribute("ownerId", taskId);
-            model.addAttribute("creatorId", userId);
+            model.addAttribute("creatorId", user.getId());
             return "comment-add-form";
-        } catch (InvalidOperation invalidOperation) {
-            invalidOperation.printStackTrace();
-            return "redirect:/task/all";
+        } catch (InvalidOperation operation) {
+            operation.printStackTrace();
         }
+        return "redirect:/";
     }
 
     @PostMapping("/add")
     public String addComment(Comment comment, long taskId, long userId) {
         try {
-            comment.setCreator(userRepository.findById(userId).orElseThrow(()->new InvalidOperation("User not found!")));
-            comment.setTask(taskRepository.findById(taskId).orElseThrow(() -> new InvalidOperation("Task not found!")));
+            comment.setCreator(userService.getByIdOrThrow(userId));
+            comment.setTask(taskService.getByIdOrThrow(taskId));
             commentService.add(comment);
             return "redirect:/task/" + taskId;
         } catch (InvalidOperation invalidOperation) {
             invalidOperation.printStackTrace();
-            return "redirect:/comment/all";
+            return "redirect:/";
         }
     }
 
     @GetMapping("/edit/{id}")
-    public String editComment(Model model, @PathVariable(name = "id") long commentId) {
+    public String editComment(Model model, Principal principal, @PathVariable(name = "id") long commentId) {
         try {
             Comment commentToEdit = commentService.getByIdOrThrow(commentId);
-            model.addAttribute("newComment", commentToEdit);
-            model.addAttribute("ownerId", commentToEdit.getTask().getId());
-            model.addAttribute("creatorId", commentToEdit.getCreator().getId());
-            return "comment-add-form";
+            if (authorizationService.isUserEligibleToEditComment(principal, commentToEdit)) {
+                model.addAttribute("newComment", commentToEdit);
+                model.addAttribute("ownerId", commentToEdit.getTask().getId());
+                model.addAttribute("creatorId", commentToEdit.getCreator().getId());
+                return "comment-add-form";
+            } else {
+                //unauthorized access
+            }
         } catch (InvalidOperation invalidOperation) {
             invalidOperation.printStackTrace();
-            return "redirect:/comment/all";
         }
+        return "redirect:/";
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteComment(@PathVariable(name = "id") long commentId) {
+    public String deleteComment(Principal principal, @PathVariable(name = "id") long commentId) {
         try {
             Comment comment = commentService.getByIdOrThrow(commentId);
-            long taskId = comment.getTask().getId();
-            commentService.delete(commentId);
-            return "redirect:/task/" + taskId;
+            if (authorizationService.isUserEligibleToDeleteComment(principal, comment)) {
+                long taskId = comment.getTask().getId();
+                commentService.delete(commentId);
+                return "redirect:/task/" + taskId;
+            } else {
+                //unauthorized access
+            }
         } catch (InvalidOperation invalidOperation) {
             invalidOperation.printStackTrace();
-            return "redirect:/task/all";
         }
+        return "redirect:/";
     }
 }
